hardware: pcb: reduce ground fill, loosen up routing
Reduce the ground fill under the IC, to prevent it from sliding around.
This gives less copper, but it should be fine for the currents we're
drawing.

Loosen up the routing of the 5V plane, which involves moving some
components around.  This increases the amount of copper that goes to the
various regulators.

Finally, reorder the caps so that the larger ones are further from the
IC.  This is done because they have a slower response time, and so can
be further away.

Signed-off-by: Sean Cross <sean@xobs.io>