hardware: sch: remove excess decoupling caps
We probably don't need 10nF, because this won't do any very-high-speed
(>100 MHz) operations.  We can think about putting them back in later,
space permitting.

Additionally, remove an extra bank of caps for one of the IO pads.
We'll double-up on capacitors there, which should be alright given the
close proximity and the fact that the only thing on that IO bank is
captouch.

Signed-off-by: Sean Cross <sean@xobs.io>