sch: mark VCCPLL regulator as DNP, use filter circuit
The VCCPLL line is extraordinarily sensitive to voltage rise times, and
appears to cause the FPGA to go into latchup very very easily.

As a result, the VCCPLL regulator burns all 250 mA of its budget
constantly, as the FPGA shunts VCCPLL to GND.  This would be ideal
during an ESD event, but not ideal during normal operations.

Indicate the VCCPLL Regulator approach is DNP, and that the VCCPLL
Filter Network is preferred.  Mostly because the regulator doesn't work
at all.

Signed-off-by: Sean Cross <sean@xobs.io>