pcb: correct RGB LED pinout, route A5 -> 3.3V
Route pin A5 to 3.3V instead of GND.  As part of the schematic change,
two capacitors needed to be rotated, though this shouldn't affect
anything electrically.

Also update the pinout for the RGB LED, but don't change the gerber
traces at all.

Practically, the only change is on layer 3 where a 3.3V pour has been
extended so that it is hit by pin A5.

Signed-off-by: Sean Cross <sean@xobs.io>